+++ screens/SearchTabs.js
@@ -1,0 +1,47 @@
+import * as React from 'react'
+import { View, StyleSheet, Dimensions, Text } from 'react-native'
+import { TabView, SceneMap } from 'react-native-tab-view'
+
+const FirstRoute = () => (
+  <View style={[styles.scene, { backgroundColor: '#ff4081' }]} />
+)
+
+const SecondRoute = () => (
+  <View style={[styles.scene, { backgroundColor: '#673ab7' }]} />
+)
+
+export default class SearchTabs extends React.Component {
+  constructor(props) {
+    super(props)
+    this.state = {
+      index: 0,
+      routes: [
+        { key: 'first', title: 'First' },
+        { key: 'second', title: 'Second' },
+      ],
+    }
+  }
+
+  render() {
+    return (
+      <View>
+        <Text>Render</Text>
+        <TabView
+          navigationState={this.state}
+          renderScene={SceneMap({
+            first: FirstRoute,
+            second: SecondRoute,
+          })}
+          onIndexChange={index => this.setState({ index })}
+          initialLayout={{ width: Dimensions.get('window').width }}
+        />
+      </View>
+    )
+  }
+}
+
+const styles = StyleSheet.create({
+  scene: {
+    flex: 1,
+  },
+})
